refactor(type): 타입 선언 파일 분리
