+++ src/component/reusable/RestaurantTicket.ts
@@ -1,14 +1,4 @@
-type Category = "한식" | "중식" | "일식" | "양식" | "아시안" | "기타";
-type TakingTime = 5 | 10 | 15 | 20 | 25 | 30;
-
-interface Restaurant {
-  category: Category;
-  name: string;
-  takingTime: TakingTime;
-  description?: string;
-  link?: string;
-  src: string;
-}
+import { Restaurant } from "../../type/type";
 
 class RestaurantTicket {
   restaurant: Restaurant;
