+++ src/domain/RestaurantController.ts
@@ -1,0 +1,3 @@
+class RestaurantController {}
+
+export default RestaurantController;
